requirements.txt
beautifulsoup4
configparser
argparse
requests
progressbar2
